Add itstool_batch_dirs option to translate entire dirs at once
Without this option, this plugin translates each file on demand. With
this option, the first time a translated page for a language in a
directory is asked for, all pages in that directory are translated to
that language in one command.

This is considerably faster (up to 50% total speed improvements in
some tests), but could be prone to error in some corner cases. So
it defaults to off for now.